+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -14,7 +14,7 @@ import java.util.Set;
 @NonNull
 public class User {
     @Positive
-    private Integer id;
+    private Long id;
     @Email(message = "Поле email не соответствует формату userEmail@email.com")
     private final String email;
     @Pattern(message = "Поле login должно содержать только A-Z и 1-0",
@@ -25,5 +25,5 @@ public class User {
     private String name = "";
     @PastOrPresent(message = "Поле birthday не корректно")
     private final LocalDate birthday;
-    private final Set<Integer> friends = new HashSet<>();
+    private final Set<Long> friends = new HashSet<>();
 }
