+++ src/main/java/ru/yandex/practicum/filmorate/storage/implement/InMemoryUserStorage.java
@@ -12,8 +12,8 @@ import java.util.*;
 @Component
 @Slf4j
 public class InMemoryUserStorage implements UserStorage {
-    private int currencyUserId = 1;
-    private final Map<Integer, User> users = new HashMap<>();
+    private long currencyUserId = 1;
+    private final Map<Long, User> users = new HashMap<>();
     private final Set<String> usageEmail = new HashSet<>();
 
     @Override
@@ -53,7 +53,7 @@ public class InMemoryUserStorage implements UserStorage {
     }
 
     @Override
-    public void deleteUser(Integer id) {
+    public void deleteUser(Long id) {
         if (!checkUser(id)) {
             throw new NotFoundException("Пользователь не обнаружен id " + id);
         }
@@ -62,7 +62,7 @@ public class InMemoryUserStorage implements UserStorage {
     }
 
     @Override
-    public User getUserById(Integer id) {
+    public User getUserById(Long id) {
         if (!checkUser(id)) {
             throw new NotFoundException("Пользователь не обнаружен id " + id);
         }
@@ -71,7 +71,7 @@ public class InMemoryUserStorage implements UserStorage {
     }
 
     @Override
-    public boolean checkUser (Integer id) {
+    public boolean checkUser (Long id) {
         log.trace("check user id '{}'", id);
         return users.containsKey(id);
     }
